move load_data functionality back in utils
